MNT-20203 Improper Output Neutralization for Logs CWE ID 117
A function call could result in a log forging attack. Writing untrusted data into a log file allows an attacker to forge log entries or inject malicious content into log files. Corrupted log files can be used to cover an attacker's tracks or as a delivery mechanism for an attack on a log viewing or processing utility. For example, if a web administrator uses a browser-based utility to review logs, a cross-site scripting attack might be possible.

- added a new java class to neutralize logs output
- imported esapi library to encode logs for html output
- override log4j to use new custom layout class as the layout pattern for share.log