MNT-20199 Improper Output Neutralization for Logs CWE ID 117
A function call could result in a log forging attack. Writing untrusted data into a log file allows an attacker to forge log entries or inject malicious content into log files. Corrupted log files can be used to cover an attacker's tracks or as a delivery mechanism for an attack on a log viewing or processing utility. For example, if a web administrator uses a browser-based utility to review logs, a cross-site scripting attack might be possible.

- imported ESAPI library
- added ESAPI.properties for configuration
- added a java class that extends PatternLayout to sanitize log message