MNT-20200 Improper Neutralization of Script-Related HTML Tags in a Web Page (Basic XSS) CWE ID 80
This call contains a cross-site scripting (XSS) flaw. The application populates the HTTP response with untrusted input, allowing an attacker to embed malicious content, such as Javascript code, which will be executed in the context of the victim's browser. Use contextual escaping on all untrusted data before using it to construct any portion of an HTTP response.

- imported OWASP Java html sanitizer library
- sanitized redirectURL in KerberosAuthenticationFilter and NTLMAuthenticationFilter classes

NOTE: this fix is meant for Alfresco 6.0 according to task description (Alfresco 6.1 does not have NTLM classes)